Aoc gcp backup/restore revisions
This commit updates the backup/restore classes and test classes to
follow the same model/standard as aoc aws backup/restore.
