+++ pytest.ini
@@ -7,6 +7,7 @@ markers =
     aoc_aws_backup_restore
     aoc_gcp_backup
     aoc_gcp_restore
+    aoc_aws_backup_restore
     gcp
     operations
 filterwarnings =
